requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
openai==1.3.7
requests==2.31.0
beautifulsoup4==4.12.2
streamlit==1.28.1
plotly==5.18.0
pandas==2.1.3
httpx==0.25.2
playwright==1.40.0
python-dateutil==2.8.2

